rename eventstream runtime
