+++ openhands/runtime/__init__.py
@@ -1,8 +1,8 @@
 from openhands.core.logger import openhands_logger as logger
-from openhands.runtime.impl.e2b.sandbox import E2BBox
-from openhands.runtime.impl.eventstream.eventstream_runtime import (
-    EventStreamRuntime,
+from openhands.runtime.impl.docker.docker_runtime import (
+    DockerRuntime,
 )
+from openhands.runtime.impl.e2b.sandbox import E2BBox
 from openhands.runtime.impl.modal.modal_runtime import ModalRuntime
 from openhands.runtime.impl.remote.remote_runtime import RemoteRuntime
 from openhands.runtime.impl.runloop.runloop_runtime import RunloopRuntime
@@ -10,8 +10,8 @@ from openhands.runtime.impl.runloop.runloop_runtime import RunloopRuntime
 
 def get_runtime_cls(name: str):
     # Local imports to avoid circular imports
-    if name == 'eventstream':
-        return EventStreamRuntime
+    if name == 'eventstream' or name == 'docker':
+        return DockerRuntime
     elif name == 'e2b':
         return E2BBox
     elif name == 'remote':
@@ -30,6 +30,6 @@ __all__ = [
     'RemoteRuntime',
     'ModalRuntime',
     'RunloopRuntime',
-    'EventStreamRuntime',
+    'DockerRuntime',
     'get_runtime_cls',
 ]
