+++ openhands/runtime/impl/runloop/runloop_runtime.py
@@ -16,7 +16,7 @@ from openhands.core.exceptions import (
 )
 from openhands.core.logger import openhands_logger as logger
 from openhands.events import EventStream
-from openhands.runtime.impl.eventstream.eventstream_runtime import EventStreamRuntime
+from openhands.runtime.impl.docker.docker_runtime import DockerRuntime
 from openhands.runtime.plugins import PluginRequirement
 from openhands.runtime.utils.command import get_remote_startup_command
 from openhands.runtime.utils.log_streamer import LogStreamer
@@ -75,8 +75,8 @@ class RunloopLogStreamer(LogStreamer):
             self.log('error', f'Error streaming runloop logs: {e}')
 
 
-class RunloopRuntime(EventStreamRuntime):
-    """The RunloopRuntime class is an EventStreamRuntime that utilizes Runloop Devbox as a runtime environment."""
+class RunloopRuntime(DockerRuntime):
+    """The RunloopRuntime class is an DockerRuntime that utilizes Runloop Devbox as a runtime environment."""
 
     _sandbox_port: int = 4444
     _vscode_port: int = 4445
@@ -211,7 +211,7 @@ class RunloopRuntime(EventStreamRuntime):
         logger.info(f'Container started. Server url: {self.api_url}')
 
         # End Runloop connect
-        # NOTE: Copied from EventStreamRuntime
+        # NOTE: Copied from DockerRuntime
         logger.info('Waiting for client to become ready...')
         self.send_status_message('STATUS$WAITING_FOR_CLIENT')
         self._wait_until_alive()
